move method definitions outside object declaration
We were using implicit `this` to refer to the object that was being exported out of AndroidInterface for web. This object was referred in presto-dom as Android.

Instead of using `this` it made sense to declare the methods separately and call each other instead of calling using `this.methodName` since there is actually no need to use `this` other than for the fact that it was being declared inside an object